requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
PyJWT>=2.10.1
cryptography>=3.4.7
pyhanko>=0.12.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
